Matched the DataTypes in Address

+++ miniproject_day9/w2g4/src/main/java/com/w2g4/w2g4/model/City.java
@@ -15,21 +15,6 @@ import lombok.AllArgsConstructor;
 public class City {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int city_id;
+    private Long city_id;
     private String name;
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
